+++ NEWS.md
@@ -1,3 +1,9 @@
+# leunbachR 0.0.3 (2026-01-23)
+
+- Fix for LRT df calculations.
+- Fix for gamma test SE, p-value, reverse one-sided test.
+- Fix for indirect equating boundary case.
+
 # leunbachR 0.0.2 (2026-01-20)
 
 - Theta scores added to output.